attempt to fix integer conversion related to BRGEMM unroll in AMX kernel

+++ src/generator_gemm_amx_microkernel.c
@@ -58,16 +58,17 @@ void libxsmm_generator_gemm_amx_prefetch_tile_in_L2(libxsmm_generated_code*     
     unsigned int tile_cols,
     unsigned int LD,
     unsigned int base_reg,
-    unsigned int offset) {
+    long long    offset) {
   unsigned int i;
   LIBXSMM_UNUSED( i_micro_kernel_config );
 
   for (i=0; i<tile_cols; i++) {
+    /* @TODO cacht overflow in offset */
     libxsmm_x86_instruction_prefetch(io_generated_code,
         LIBXSMM_X86_INSTR_PREFETCHT1,
         base_reg,
         LIBXSMM_X86_GP_REG_UNDEF, 0,
-        offset + i * LD * 2 /*(i_micro_kernel_config->datatype_size/2)*/);
+        (int)(offset + i * LD * 2 /*(i_micro_kernel_config->datatype_size/2)*/) );
   }
 }
 
@@ -660,12 +661,13 @@ void libxsmm_generator_gemm_amx_single_tilestore( libxsmm_generated_code*       
   }
 }
 
+/* @TODO catch overflows caused by a_offs */
 LIBXSMM_API_INTERN
 void libxsmm_generator_gemm_amx_decompress_32x32_A_block(libxsmm_generated_code*     io_generated_code,
     libxsmm_loop_label_tracker*        io_loop_label_tracker,
     const libxsmm_gp_reg_mapping*      i_gp_reg_mapping,
     libxsmm_micro_kernel_config*       i_micro_kernel_config,
-    int                                a_offs,
+    long long                          a_offs,
     unsigned int                       a_lookahead_offs,
     unsigned int                       a_lookahead_br_index) {
 
@@ -717,7 +719,7 @@ void libxsmm_generator_gemm_amx_decompress_32x32_A_block(libxsmm_generated_code*
         i_gp_reg_mapping->gp_reg_bitmap_a,
         decompress_loop_reg,
         1,
-        (a_offs*i_micro_kernel_config->sparsity_factor_A)/16 + expanded_cl * 4 + (a_lookahead_offs * i_micro_kernel_config->sparsity_factor_A)/16,
+        (int)((a_offs*i_micro_kernel_config->sparsity_factor_A)/16 + expanded_cl * 4 + (a_lookahead_offs * i_micro_kernel_config->sparsity_factor_A)/16),
         current_mask_reg );
 
     /* Expand operation */
@@ -727,7 +729,7 @@ void libxsmm_generator_gemm_amx_decompress_32x32_A_block(libxsmm_generated_code*
                                                  i_gp_reg_mapping->gp_reg_a,
                                                  n_elts_decompressed_reg,
                                                  2,
-                                                 a_offs + a_lookahead_offs,
+                                                 (int)(a_offs + a_lookahead_offs),
                                                  0,
                                                  LIBXSMM_X86_VEC_REG_UNDEF,
                                                  current_zmm,
@@ -765,7 +767,7 @@ void libxsmm_generator_gemm_amx_decompress_32x32_A_block(libxsmm_generated_code*
       LIBXSMM_X86_INSTR_VMOVUPS,
       i_gp_reg_mapping->gp_reg_decompressed_a,
       decompress_loop_reg, 8,
-      a_offs * i_micro_kernel_config->sparsity_factor_A + expanded_cl * 64 + a_lookahead_offs * i_micro_kernel_config->sparsity_factor_A,
+      (int)(a_offs * i_micro_kernel_config->sparsity_factor_A + expanded_cl * 64 + a_lookahead_offs * i_micro_kernel_config->sparsity_factor_A),
       i_micro_kernel_config->vector_name,
       current_zmm, 0, 0, 1 );
   }
@@ -886,8 +888,8 @@ void libxsmm_generator_gemm_amx_microkernel( libxsmm_generated_code*            
     const libxsmm_gemm_descriptor*     i_xgemm_desc,
     libxsmm_blocking_info_t*           n_blocking_info,
     libxsmm_blocking_info_t*           m_blocking_info,
-    unsigned long long                 offset_A,
-    unsigned long long                 offset_B,
+    long long                          offset_A,
+    long long                          offset_B,
     unsigned int                       is_last_k,
     long long                          i_brgemm_loop,
     unsigned int                       fully_unrolled_brloop  ) {
@@ -938,8 +940,8 @@ void libxsmm_generator_gemm_amx_microkernel( libxsmm_generated_code*            
   int _A_tileload_instr[4] = { 0 };
   int _B_tileload_instr[4] = { 0 };
   int _C_tilecomp_instr[4] = { 0 };
-  unsigned long long _A_offsets[4] = { 0 };
-  unsigned long long _B_offsets[4] = { 0 };
+  long long _A_offsets[4] = { 0 };
+  long long _B_offsets[4] = { 0 };
   int _im_offset_prefix_sums[4] = { 0 };
   int _in_offset_prefix_sums[4] = { 0 };
 
@@ -1148,13 +1150,14 @@ void libxsmm_generator_gemm_amx_microkernel( libxsmm_generated_code*            
     }
 
     if (_A_tile_id_load[i] > 0) {
+      /* @TODO: catch int overdlow in A_offset */
       libxsmm_x86_instruction_tile_move( io_generated_code,
           i_micro_kernel_config->instruction_set,
           _A_tileload_instr[i],
           gp_reg_a,
           i_gp_reg_mapping->gp_reg_lda,
           4,
-          _A_offsets[i] * i_micro_kernel_config->sparsity_factor_A,
+          (int)(_A_offsets[i] * i_micro_kernel_config->sparsity_factor_A),
           _A_tile_id_load[i]);
 
       if (i_brgemm_loop + pf_dist < i_xgemm_desc->c3) {
@@ -1164,18 +1167,19 @@ void libxsmm_generator_gemm_amx_microkernel( libxsmm_generated_code*            
             n_CL_to_pf,
             i_xgemm_desc->lda * 2,
             i_gp_reg_mapping->gp_reg_a,
-            (unsigned int)(_A_offsets[i] + pf_dist * i_xgemm_desc->c1));
+            _A_offsets[i] + pf_dist * i_xgemm_desc->c1 );
       }
     }
 
     if (_B_tile_id_load[i] > 0) {
-        libxsmm_x86_instruction_tile_move( io_generated_code,
+      /* @TODO: catch int overdlow in B_offset */
+      libxsmm_x86_instruction_tile_move( io_generated_code,
             i_micro_kernel_config->instruction_set,
             _B_tileload_instr[i],
             i_gp_reg_mapping->gp_reg_b,
             i_gp_reg_mapping->gp_reg_ldb,
             4,
-            _B_offsets[i],
+            (int)_B_offsets[i],
             _B_tile_id_load[i]);
 
       if (i_micro_kernel_config->norm_to_normT_B_ext_buf == 1) {
@@ -1190,7 +1194,7 @@ void libxsmm_generator_gemm_amx_microkernel( libxsmm_generated_code*            
             n_CL_to_pf,
             i_xgemm_desc->ldb * 2,
             i_gp_reg_mapping->gp_reg_b,
-            (unsigned int)(_B_offsets[i] + pf_dist * i_xgemm_desc->c2));
+            _B_offsets[i] + pf_dist * i_xgemm_desc->c2 );
       }
     }
 
@@ -1243,8 +1247,8 @@ void libxsmm_generator_gemm_amx_kernel_kloop( libxsmm_generated_code*           
     const libxsmm_gemm_descriptor*     i_xgemm_desc,
     libxsmm_blocking_info_t*           n_blocking_info,
     libxsmm_blocking_info_t*           m_blocking_info,
-    unsigned long long                 A_offs,
-    unsigned long long                 B_offs,
+    long long                          A_offs,
+    long long                          B_offs,
     unsigned int                       fully_unrolled_brloop ) {
 
   unsigned int l_k_blocking = 16;
