+++ src/generator_gemm_amx_microkernel_emu.c
@@ -626,10 +626,10 @@ void libxsmm_generator_gemm_amx_microkernel_emu( libxsmm_generated_code*        
     const libxsmm_gemm_descriptor*     i_xgemm_desc,
     libxsmm_blocking_info_t*           n_blocking_info,
     libxsmm_blocking_info_t*           m_blocking_info,
-    unsigned int                       offset_A,
-    unsigned int                       offset_B,
+    long long                          offset_A,
+    long long                          offset_B,
     unsigned int                       is_last_k,
-    int                                i_brgemm_loop,
+    long long                          i_brgemm_loop,
     unsigned int                       fully_unrolled_brloop  ) {
   int m_tiles = m_blocking_info->tiles;
   int n_tiles = n_blocking_info->tiles;
@@ -678,8 +678,8 @@ void libxsmm_generator_gemm_amx_microkernel_emu( libxsmm_generated_code*        
   int _A_tileload_instr[4] = { 0 };
   int _B_tileload_instr[4] = { 0 };
   int _C_tilecomp_instr[4] = { 0 };
-  int _A_offsets[4] = { 0 };
-  int _B_offsets[4] = { 0 };
+  long long _A_offsets[4] = { 0 };
+  long long _B_offsets[4] = { 0 };
   int _im_offset_prefix_sums[4] = { 0 };
   int _in_offset_prefix_sums[4] = { 0 };
 
@@ -983,15 +983,15 @@ void libxsmm_generator_gemm_amx_kernel_kloop_emu( libxsmm_generated_code*       
     const libxsmm_gemm_descriptor*     i_xgemm_desc,
     libxsmm_blocking_info_t*           n_blocking_info,
     libxsmm_blocking_info_t*           m_blocking_info,
-    unsigned int                       A_offs,
-    unsigned int                       B_offs,
+    long long                          A_offs,
+    long long                          B_offs,
     unsigned int                       fully_unrolled_brloop ) {
 
   unsigned int l_k_blocking = 16;
   unsigned int k;
-  unsigned int offset_A = 0;
-  unsigned int offset_B = 0;
-  int i_brgemm_loop = -2;
+  long long offset_A = 0;
+  long long offset_B = 0;
+  long long i_brgemm_loop = -2;
   int is_last_k = 0;
 
   if (i_xgemm_desc->flags & LIBXSMM_GEMM_FLAG_BATCH_REDUCE_STRIDE) {
